update for third party sign in
Finally got third party working. Needed to update tables according to NextAuth documentation in the prisma schema. Then needed to update signin callback in the auth api to allow for account updating or creation on sign in. Confirmed it's working. Also added in redirect to profile on successful sign in.